Moved game code out of index.js to game.js, and created menu.js.
