+++ src/menu.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+/**
+ * Created by pc on 2/13/2017.
+ */
+export class Menu extends React.Component {
+  render() {
+    return (
+      <div className="menu">
+        <h2>Menu</h2>
+      </div>
+    )
+  }
+}
